requirements.txt
Pillow==10.3.0
streamlit==1.21.0
